refactor: add game result judge
